xml files added to SpringCRUD and TestSpring@Autowired added

+++ SpringCRUD/src/main/java/service/impl/StudentServiceImpl.java
@@ -2,12 +2,13 @@ package service.impl;
 
 import model.Student;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import repository.StudentRepository;
 import repository.impl.StudentRepositoryImpl;
 import service.StudentService;
 
 import java.util.List;
-
+@Service
 public class StudentServiceImpl implements StudentService {
     @Autowired
     private StudentRepository studentRepository;
@@ -19,27 +20,35 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void addStudent(Student student) {
-        studentRepository.addStudent(student);
+        if (student != null) {
+            studentRepository.addStudent(student);
+        } else throw new RuntimeException("student is null");
+
     }
 
     @Override
     public void deleteStudent(Integer id) {
-        studentRepository.deleteStudent(id);
+        if (id != null) {
+            studentRepository.deleteStudent(id);
+        } else throw new RuntimeException("id is null");
     }
 
     @Override
     public void updateStudent(Student student) {
+        if (student != null) {
         studentRepository.updateStudent(student);
+        } else throw new RuntimeException("student is null");
     }
 
     @Override
     public Student getStudent(Integer id) {
-
+        if (id != null) {
         return studentRepository.getStudent(id);
+        } else throw new RuntimeException("id is null");
     }
 
     @Override
     public List<Student> getAllStudents() {
-        return studentRepository.getAllStudents();
+        return (List<Student>)studentRepository.getAllStudents();
     }
 }
